+++ src/rogallo/data/__init__.py
@@ -26,9 +26,12 @@ from .state import (
     load_command_history,
     load_location_history,
     load_navigation_history,
+    load_ui_state,
     save_command_history,
     save_location_history,
     save_naviagation_history,
+    save_ui_state,
+    update_ui_state,
 )
 from .trust import trust_file
 from .trusted_mime_types import load_trusted_mime_types, save_trusted_mime_types
@@ -56,6 +59,7 @@ __all__ = [
     "load_toolbar",
     "load_trusted_mime_types",
     "load_trusted_schemes",
+    "load_ui_state",
     "LocationHistory",
     "LocationVisit",
     "NavigationHistory",
@@ -67,8 +71,10 @@ __all__ = [
     "save_naviagation_history",
     "save_trusted_mime_types",
     "save_trusted_schemes",
+    "save_ui_state",
     "trust_file",
     "update_configuration",
+    "update_ui_state",
 ]
 
 ### __init__.py ends here