+++ src/rogallo/screens/main/screen.py
@@ -71,11 +71,7 @@ from ...commands import (
     SearchHistory,
     SetHome,
     SetHomeToCurrentLocation,
-    StripeLinks,
     ToggleANSIEscapeSequenceHandling,
-    ToggleCosyLinkNumbers,
-    ToggleEmojiRemoval,
-    ToggleLinkNumbers,
     ToggleSidePanel,
     ToggleView,
     ViewChangeLog,
@@ -96,11 +92,13 @@ from ...data import (
     load_toolbar,
     load_trusted_mime_types,
     load_trusted_schemes,
+    load_ui_state,
     save_bookmarks,
     save_command_history,
     save_location_history,
     save_naviagation_history,
     update_configuration,
+    update_ui_state,
 )
 from ...input_content import InputContent
 from ...messages import (
@@ -241,11 +239,7 @@ class Main(EnhancedScreen[None]):
         SaveSource,
         SetHome,
         SetHomeToCurrentLocation,
-        StripeLinks,
         ToggleANSIEscapeSequenceHandling,
-        ToggleCosyLinkNumbers,
-        ToggleEmojiRemoval,
-        ToggleLinkNumbers,
         ToggleSidePanel,
         ToggleView,
         ViewChangeLog,
@@ -314,7 +308,7 @@ class Main(EnhancedScreen[None]):
 
     def _watch__side_panel_visible(self) -> None:
         """Watch for changes to the side panel visibility."""
-        with update_configuration() as config:
+        with update_ui_state() as config:
             config.side_panel_visible = self._side_panel_visible
 
     def compose(self) -> ComposeResult:
@@ -359,8 +353,7 @@ class Main(EnhancedScreen[None]):
             await self._clients.gemini.client_cert_store.list_certificates()
         )
         config = load_configuration()
-        self._side_panel_visible = config.side_panel_visible
-        self._side_panel.dock_right = config.side_panel_on_right
+        self._side_panel_visible = load_ui_state().side_panel_visible
         self._command_line.dock_top = config.command_line_on_top
         if self._clients.gemini.trust_store:
             self._command_line.known_hosts = [
@@ -368,11 +361,7 @@ class Main(EnhancedScreen[None]):
                 for host, port in await self._clients.gemini.trust_store.get_hosts()
             ]
             HistorySearchCommands.known_hosts = self._command_line.known_hosts
-        self._viewer.stripe_links = config.stripe_links
-        self._viewer.with_link_numbers = config.with_link_jumps
         self._viewer.handle_ansi_escape_sequences = config.handle_ansi_escape_sequences
-        self._viewer.strip_emoji = config.strip_emoji
-        self._viewer.cosy_link_numbers = config.cosy_link_jumps
         if self._arguments.command == "open" and (
             location := getattr(self._arguments, "location", None)
         ):
@@ -869,24 +858,6 @@ class Main(EnhancedScreen[None]):
             self._cache.clear()
             self.notify("All cached content has been cleared.", title="Cache")
 
-    def action_stripe_links_command(self) -> None:
-        """Toggle link striping."""
-        self._viewer.stripe_links = not self._viewer.stripe_links
-        with update_configuration() as config:
-            config.stripe_links = self._viewer.stripe_links
-
-    def action_toggle_link_numbers_command(self) -> None:
-        """Toggle link numbers."""
-        self._viewer.with_link_numbers = not self._viewer.with_link_numbers
-        with update_configuration() as config:
-            config.with_link_jumps = self._viewer.with_link_numbers
-
-    def action_toggle_cosy_link_numbers_command(self) -> None:
-        """Toggle cosy link numbers."""
-        self._viewer.cosy_link_numbers = not self._viewer.cosy_link_numbers
-        with update_configuration() as config:
-            config.cosy_link_jumps = self._viewer.cosy_link_numbers
-
     def action_go_to_parent_command(self) -> None:
         """Go to the parent of the current document's location."""
         if (
@@ -902,12 +873,6 @@ class Main(EnhancedScreen[None]):
             and location.root != location
         ):
             self.post_message(OpenLocation(location.root))
-
-    def action_toggle_emoji_removal_command(self) -> None:
-        """Toggle emoji removal."""
-        self._viewer.strip_emoji = not self._viewer.strip_emoji
-        with update_configuration() as config:
-            config.strip_emoji = self._viewer.strip_emoji
 
     def action_toggle_ansi_escape_sequence_handling_command(self) -> None:
         """Toggle ANSI escape sequence handling."""